refactor(collectors): rework dispatching logic used by collectors
The dispatching logic lives in a dedicated function and each
collector calls it in their Start function.
A lot of duplicated code is removed.
