+++ collectors/replicaset.go
@@ -16,7 +16,6 @@ package collectors
 
 import (
 	"context"
-	"sync"
 
 	"github.com/go-logr/logr"
 	v1 "k8s.io/api/apps/v1"
@@ -143,43 +142,7 @@ func (r *ReplicasetCollector) Reconcile(ctx context.Context, req ctrl.Request) (
 // using the manager. It starts go routines needed by the collector to interact with the
 // broker.
 func (r *ReplicasetCollector) Start(ctx context.Context) error {
-	wg := sync.WaitGroup{}
-	// it listens for new subscribers and sends the cached events to the
-	// subscriber received on the channel.
-	dispatchEventsOnSubcribe := func(ctx context.Context) {
-		wg.Add(1)
-		for {
-			select {
-			case sub := <-r.SubscriberChan:
-				r.logger.V(5).Info("Dispatching events", "subscriber", sub)
-				dispatch := func(res *events.GenericResource) {
-					// Check if the pod is related to the subscriber.
-					if _, ok := res.Nodes[sub]; ok {
-						r.Queue.Push(res.ToEvent(events.Added, []string{sub}))
-					}
-				}
-				r.Cache.ForEach(dispatch)
-
-			case <-ctx.Done():
-				r.logger.V(5).Info("Stopping dispatcher on new subscribers")
-				wg.Done()
-				return
-			}
-		}
-	}
-
-	r.logger.Info("Starting event dispatcher for new subscribers")
-	// Start the dispatcher.
-	go dispatchEventsOnSubcribe(ctx)
-
-	// Wait for shutdown signal.
-	<-ctx.Done()
-	r.logger.Info("Waiting for event dispatcher to finish")
-	// Wait for goroutines to stop.
-	wg.Wait()
-	r.logger.Info("Dispatcher finished")
-
-	return nil
+	return dispatch(ctx, r.logger, r.SubscriberChan, r.Queue, &r.Cache)
 }
 
 func (r *ReplicasetCollector) initMetrics() {
